Issue #20, add removeEntry and removeEntries methods.
Removing is done on the fly, without unpacking, deleting entries and
packing back. This adds some complexity and performance overhead with
handling directories correctly. 

Current implementation scans given entries names, determines which are
directories and then tests every entry against that list to determine if
current entry is inside of one of the dirs to remove.