Timestamp preserving tests get custom handling
The timestamp preserving tests don't work with Java 8 deterministicly the way they are written now. They
use the old API of setTime and getTime which do some DOS time mangling that results in some seconds
getting shaven off.

For the Java8 API we support setting the modified time, last access time and creation time. If I add those
tests then we need to start building the project only with Java8. We'll have a discussion about it.